+++ src/ENDFtk/section/Base/src/ctor.hpp
@@ -6,9 +6,11 @@
 Base() = default;
 #endif
 
-Base( double ZA, double AWR, int MT ) :
+Base( int ZA, double AWR, int MT ) :
   BaseWithoutMT( ZA, AWR ), MT_( MT ) {
-  if ( MT < 1 ){
+
+  if ( MT < 1 ) {
+
     Log::error( "Illegal section number" );
     Log::info( "Section (MT) numbers must be greater than zero" );
     Log::info( "Encountered section number: {}", MT );
@@ -18,13 +20,16 @@ Base( double ZA, double AWR, int MT ) :
 
 Base( const HEAD& head, int MAT, int MF ) :
   Base( head.ZA(), head.atomicWeightRatio(), head.MT() ) {
-  if( MAT != head.MAT() ) {
+    
+  if ( MAT != head.MAT() ) {
+
     Log::error( "Incorrect Material number (MAT)." );
     Log::info( "Expected MAT = {}", MAT );
     Log::info( "Encountered MAT = {}", head.MAT() );
     throw std::exception();
   }
-  if( MF != head.MF() ) {
+  if ( MF != head.MF() ) {
+
     Log::error( "Incorrect File number (MF)." );
     Log::info( "Expected MF = {}", MF );
     Log::info( "Encountered MF = {}", head.MF() );
